Add windows XP support
I was trying to use "start-selenium" (cool project !!) in windows xp environment but it fails with error like "could not find java... ", but java was correctly installed. So following the code I found that there is an issue with the "whereis" module because "where" command is not valid in windows xp.